x/crowdfund/types/message_pledge_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"github.com/youngestdev/crowdfund/testutil/sample"
)

func TestMsgPledge_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgPledge
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgPledge{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgPledge{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
